Add a copy of the code currently running on the FUT server
BACKGROUND:
According to the original developer of the GUI, the code that was originally checked into this repository isn't completely up to date. We need a copy of the code that's actually running on FUT. This commit copies all that over, with only one caveat.

CAVEAT:
In the code running on FUT, a lot of the files are owned by `root` for some reason. I had to change their ownership to download them. So the only difference between the code I'm uploading here and the code in FUT should be that.