cleanup: run autoformatters over our codebase
BACKGROUND:
Our codebase sometimes doesn't follow standard style for PHP, HTML,
CSS, or JavaScript. It would be nice to make our code do that, because
then it'll be more readable.

This commit runs autoformatters over our codebase. Specifically, it runs
VS Code's built-in autoformatter for HTML, CSS, and JavaScript, and it
runs DEVSENSE's autoformatter for PHP.