fix: stop jumping to page top when deleting router
BACKGROUND:
Right now, when you delete a router, your browser window will
automatically jump to the top of the page. That's annoying and bad, so
we should make that stop happening. This commit does that, and it also
stops the window from jumping when you edit a router in general.